autocfs-client project update 2018-04-02 10:16

+++ autocfs-core/src/main/java/com/autocfs/io/Copier.java
@@ -3,5 +3,5 @@ package com.autocfs.io;
 import java.io.IOException;
 
 public interface Copier {
-    public int transfer() throws IOException;
+    public long transfer() throws IOException;
 }
